+++ arch/s390/include/asm/vdso/gettimeofday.h
@@ -20,20 +20,20 @@ static inline u64 __arch_get_hw_counter(s32 clock_mode, const struct vdso_time_d
 }
 
 static __always_inline
-long clock_gettime_fallback(clockid_t clkid, struct __kernel_timespec *ts)
+int clock_gettime_fallback(clockid_t clkid, struct __kernel_timespec *ts)
 {
 	return syscall2(__NR_clock_gettime, (long)clkid, (long)ts);
 }
 
 static __always_inline
-long gettimeofday_fallback(register struct __kernel_old_timeval *tv,
-			   register struct timezone *tz)
+int gettimeofday_fallback(register struct __kernel_old_timeval *tv,
+			  register struct timezone *tz)
 {
 	return syscall2(__NR_gettimeofday, (long)tv, (long)tz);
 }
 
 static __always_inline
-long clock_getres_fallback(clockid_t clkid, struct __kernel_timespec *ts)
+int clock_getres_fallback(clockid_t clkid, struct __kernel_timespec *ts)
 {
 	return syscall2(__NR_clock_getres, (long)clkid, (long)ts);
 }
